% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plane3D.R
\name{plane3D}
\alias{plane3D}
\title{plane3D}
\usage{
plane3D(
  z,
  x,
  y,
  zlab = NULL,
  xlab = NULL,
  ylab = NULL,
  zlim = NULL,
  xlim = NULL,
  ylim = NULL,
  cis = TRUE,
  nlines = 5,
  heatmap = NULL,
  heatmap_cols = NULL,
  heatmap_border = NA,
  plane0 = FALSE,
  plane_aux = FALSE,
  plane_aux_pos = NULL,
  plane_aux_col = NULL,
  main = NULL,
  cex.main = 1,
  nticks = NULL,
  expand = 0.8,
  theta = 45,
  phi = 0,
  r = sqrt(3),
  d = 1.2
)
}
\arguments{
\item{z}{A two-dimensional matrix of point estimates or a three-dimensional
array of of point estimates and interval values.}

\item{x}{A vector of \code{x}-values, matching the first dimension of
\code{z}.}

\item{y}{A vector of \code{y}-values, matching the second dimension of
\code{z}.}

\item{zlab}{titles for the axes.  N.B. These must be
    character strings; expressions are not accepted.  Numbers will be
    coerced to character strings.}

\item{xlab}{titles for the axes.  N.B. These must be
    character strings; expressions are not accepted.  Numbers will be
    coerced to character strings.}

\item{ylab}{titles for the axes.  N.B. These must be
    character strings; expressions are not accepted.  Numbers will be
    coerced to character strings.}

\item{zlim}{x-, y-  and z-limits.  These should be chosen
    to cover the range of values of the surface: see \sQuote{Details}.}

\item{xlim}{x-, y-  and z-limits.  These should be chosen
    to cover the range of values of the surface: see \sQuote{Details}.}

\item{ylim}{x-, y-  and z-limits.  These should be chosen
    to cover the range of values of the surface: see \sQuote{Details}.}

\item{cis}{Plot confidence intervals? Defaults to \code{TRUE}.}

\item{nlines}{Suggested number of lines for the grid display of the
regression plane. May be overridden.}

\item{heatmap}{A two-dimensional matrix indicating the joint frequency
distribution of \code{x} and \code{y}. Will displayed as a shaded histogram
below the plot.}

\item{heatmap_cols}{Override default shadings of \code{heatmap}.}

\item{heatmap_border}{Override default border color of \code{heatmap}.}

\item{plane0}{Add a a horizontal red plane at \code{z == 0}.
Defaults to \code{FALSE}.}

\item{plane_aux}{Add an auxiliary reference place. Defaults to \code{FALSE}.}

\item{plane_aux_pos}{User-specified position of auxiliary reference plane if
\code{plane_aux == TRUE}.}

\item{plane_aux_col}{User-specified color of auxiliary reference plane if
\code{plane_aux == TRUE}.}

\item{main}{main and sub title, as for \code{\link[graphics]{title}}.}

\item{cex.main}{The magnification to be used for main titles relative to the
current setting of cex as in \code{\link[graphics]{par}}.}

\item{nticks}{the (approximate) number of tick marks to draw on the
    axes.  Has no effect if \code{ticktype} is \code{"simple"}.}

\item{expand}{a expansion factor applied to the \code{z}
    coordinates. Often used with \code{0 < expand < 1} to shrink the
    plotting box in the \code{z} direction.}

\item{theta}{angles defining the viewing direction.
    \code{theta} gives the azimuthal direction and \code{phi}
    the colatitude.}

\item{phi}{angles defining the viewing direction.
    \code{theta} gives the azimuthal direction and \code{phi}
    the colatitude.}

\item{r}{the distance of the eyepoint from the centre of the plotting box.}

\item{d}{a value which can be used to vary the strength of
    the perspective transformation.  Values of \code{d} greater
    than 1 will lessen the perspective effect and values less
    and 1 will exaggerate it.}
}
\value{
Returns, as invisible, the viewing transformation matrix.
}
\description{
Plot a three-dimensional regression prediction with confidence
intervals.
}
\examples{
## ---- Estimation ----
mod <- lm(vote ~
            growth +
            approval,
          dat = us)

## ---- Prediction ----
growth_range <- c(-4, 6)
approval_range <- c(30, 80)
growth_seq <-
  seq(growth_range[1], growth_range[2], length.out = 21L)
approval_seq <-
  seq(approval_range[1], approval_range[2], length.out = 21L)

pred <- array(NA, dim = c(21L, 21L, 3L))
for (growth in seq_along(growth_seq)) {
  for (approval in seq_along(approval_seq)) {
    pred_tmp <- predict.lm(
      mod,
      newdata = data.frame(growth = growth_seq[growth],
                           approval = approval_seq[approval]),
      se.fit = TRUE
    )
    pred[growth, approval,] <- c(
      pred_tmp$fit,
      pred_tmp$fit + qnorm(.025) * pred_tmp$se.fit,
      pred_tmp$fit + qnorm(.975) * pred_tmp$se.fit
    )
  }
}

## ---- Plot ----
plane3D(
  pred,
  growth_seq,
  approval_seq,
  zlab = "Predicted Vote Share",
  xlab = "Economic Growth",
  ylab = "Approval Rating",
  zlim = c(35, 70),
  xlim = growth_range,
  ylim = approval_range,
  cis = TRUE,
  nlines = 7,
  main = "Incumbent Vote Shares, Economic \n Growth, and Approval Ratings",
  theta = -45,
  phi = 9
)
}
\references{
\url{https://www.mzes.uni-mannheim.de/socialsciencedatalab/article/regplane3D/}
}
\seealso{
\code{\link{plot3D}}
}
