% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pretty_axis_inputs.R
\name{pretty_axis_inputs}
\alias{pretty_axis_inputs}
\title{determine_axis_inputs}
\usage{
pretty_axis_inputs(axis_range, base = 1, nlines_suggest = 5L, multiply = 3)
}
\arguments{
\item{axis_range}{Value range of the variable displayed on either the
\code{x}-axis or \code{y}-axis of a \code{regplane3D} plot.}

\item{base}{Numerical base to which \code{axis_range} will be rounded for
pretty display.}

\item{nlines_suggest}{Suggested number of lines for the grid display of the
regression plane. May be adjusted by the function to ensure pretty display.}

\item{multiply}{Multiply}
}
\value{
Returns a list with three entries: \code{nlines}, \code{range},
and
}
\description{
Determine axis inputs for \code{regplane3D} plotting functions.
The function returns a rounded range for the axis (\code{range}),
(possibly) an adjusted number of lines that ensure pretty display
(\code{nlines}), and a value sequence for predictions that includes the values
at which \code{z}-values for each line will be calculated and (if
\code{multiply > 1}) equally space gradations inbetween (\code{seq}).
}
\examples{
## Find range of variable
growth_range <- range(us$growth)
growth_range

## Determine axis inputs
growth_axis <- pretty_axis_inputs(
  axis_range = growth_range,
  base = 2,
  nlines_suggest = 6L,
  multiply = 4
)
growth_axis
}
\references{
\url{https://www.mzes.uni-mannheim.de/socialsciencedatalab/article/regplane3D/}
}
\seealso{
\code{\link[base]{pretty}}
}
