% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regplane3D.R
\name{regplane3D}
\alias{regplane3D}
\title{regplane3D}
\usage{
regplane3D(
  z,
  x,
  y,
  zlab = NULL,
  xlab = NULL,
  ylab = NULL,
  zlim = NULL,
  xlim = NULL,
  ylim = NULL,
  cis = TRUE,
  nlines = 5,
  heatmap = NULL,
  heatmap_cols = NA,
  heatmap_border = NA,
  plane0 = FALSE,
  plane_aux = FALSE,
  plane_aux_pos = NULL,
  plane_aux_col = NULL,
  main = NULL,
  cex.main = 1,
  nticks = NULL,
  expand = 0.8,
  theta = 45,
  phi = 0,
  r = sqrt(3),
  d = 1.2
)
}
\arguments{
\item{z}{A two-dimensional matrix of point estimates or a three-dimensional
array of of point estimates and interval values.}

\item{x}{A vector of \code{x}-values, matching the first dimension of
\code{z}.}

\item{y}{A vector of \code{y}-values, matching the second dimension of
\code{z}.}

\item{zlab}{Label for the vertical \code{z}-axis.}

\item{xlab}{Label for the horizontal \code{x}-axis.}

\item{ylab}{Label for the horizontal \code{y}-axis.}

\item{cis}{Plot confidence intervals? Defaults to \code{TRUE}.}

\item{nlines}{Suggested number of lines for the grid display of the
regression plane. May be overridden.}

\item{heatmap}{A two-dimensional matrix indicating the joint frequency
distribution of \code{x} and \code{y}. Will displayed as a shaded histogram
below the plot.}

\item{heatmap_cols}{Override default shadings of \code{heatmap}.}

\item{heatmap_border}{Override default border color of \code{heatmap}.}

\item{plane0}{Add a a horizontal red plane at \code{z == 0}.
Defaults to \code{FALSE}.}

\item{plane_aux}{Add an auxiliary reference place. Defaults to \code{FALSE}.}

\item{plane_aux_pos}{User-specified position of auxiliary reference plane if
\code{plane_aux == TRUE}.}

\item{plane_aux_col}{User-specified color of auxiliary reference plane if
\code{plane_aux == TRUE}.}

\item{...}{Additional arguments passed to the plotting methods. See
\code{\link{persp3D}} and \code{\link{persp}}.}
}
\value{
Returns, as invisible, the viewing transformation matrix.
}
\description{
Plot a three-dimensional regression prediction with confidence
intervals.
}
\examples{
## ---- Estimation ----
mod <- lm(vote ~
            growth +
            approval,
          dat = us)

## ---- Prediction ----
growth_range <- c(-4, 6)
approval_range <- c(30, 80)
growth_seq <-
  seq(growth_range[1], growth_range[2], length.out = 21L)
approval_seq <-
  seq(approval_range[1], approval_range[2], length.out = 21L)

pred <- array(NA, dim = c(21L, 21L, 3L))
for (growth in seq_along(growth_seq)) {
  for (approval in seq_along(approval_seq)) {
    pred_tmp <- predict.lm(
      mod,
      newdata = data.frame(growth = growth_seq[growth],
                           approval = approval_seq[approval]),
      se.fit = TRUE
    )
    pred[growth, approval,] <- c(
      pred_tmp$fit,
      pred_tmp$fit + qnorm(.025) * pred_tmp$se.fit,
      pred_tmp$fit + qnorm(.975) * pred_tmp$se.fit
    )
  }
}

## ---- Plot ----
regplane3D(
  pred,
  growth_seq,
  approval_seq,
  zlab = "Predicted Vote Share",
  xlab = "Economic Growth",
  ylab = "Approval Rating",
  zlim = c(35, 70),
  xlim = growth_range,
  ylim = approval_range,
  cis = TRUE,
  nlines = 7,
  main = "Incumbent Vote Shares, Economic \n Growth, and Approval Ratings",
  theta = -45,
  phi = 9
)
}
\references{
\url{https://www.mzes.uni-mannheim.de/socialsciencedatalab/article/regplane3D/}
}
\seealso{
\code{\link{plot3D}}
}
