% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/us.R
\docType{data}
\name{us}
\alias{us}
\title{US Presidential Elections 1948-2004}
\format{
An object of class \code{"tbl_df"}. Contains 8 variables for 15 observations.
\describe{
 \item{year}{Election year}
 \item{vote}{Incumbent vote share in percent}
 \item{party}{Incumbent's party}
 \item{incumbent}{Name of the incumbent president}
 \item{approval}{Incumbent's approval rating}
 \item{approval_above_mean}{A binary indicator that equals 1 if the approval rating is above the sample average}
 \item{centered_approval}{A mean-centered version of incumbent's approval rating}
}
}
\usage{
data(us)
}
\description{
A small data set containing information of incumbent vote shares,
approval ratings, and economic growth rates in US Presidential Elections
1948-2004.
}
\examples{
data(us)
head(us)

}
\keyword{datasets}
