% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap3D.R
\name{heatmap3D}
\alias{heatmap3D}
\title{heatmap3D}
\usage{
heatmap3D(
  z,
  x,
  y,
  zlab = NULL,
  xlab = NULL,
  ylab = NULL,
  zlim = NULL,
  xlim = NULL,
  ylim = NULL,
  nlines = 5,
  cols = NULL,
  border = NA,
  main = NULL,
  cex.main = 1,
  nticks = NULL,
  expand = 0.8,
  theta = 45,
  phi = 0,
  r = sqrt(3),
  d = 1.2,
  add = FALSE
)
}
\arguments{
\item{z}{A two-dimensional matrix of the discretized density or frequency
of \code{x} and \code{y}.}

\item{x}{A numeric vector of length \code{nrow(z) + 1}.}

\item{y}{A numeric vector of length \code{ncol(z) + 1}.}

\item{zlab}{Label for the vertical \code{z}-axis}

\item{xlab}{Label for the horizontal \code{x}-axis}

\item{ylab}{Label for the horizontal \code{y}-axis}

\item{nlines}{Suggested number of lines for the grid. May be overridden.}

\item{cols}{Color palette for the heatmap shadings.}

\item{...}{Additional arguments passed to the plotting methods. See
\code{\link{persp3D}} and \code{\link{hist3D}}.}
}
\value{
Returns, as invisible, the viewing transformation matrix.
}
\description{
Add a three-dimensional histogram/heatmap
}
\seealso{
\code{\link{plot3D}}
}
