% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regplane3D.R
\docType{package}
\name{regplane3D}
\alias{regplane3D}
\title{regplane3D: Plotting regression predictions in 3D}
\description{
#' @section Description:
The `regplane3D` package is a convenience wrapper for the
[`plot3D`](https://cran.r-project.org/web/packages/plot3D/index.html) package.
`regplane3D` uses several `plot3D` functions to produce visually appealing
three-dimensional displays of regression estimates with confidence intervals.
}
\section{Functions}{


`regplane3D` (development version 0.1.0) consists of three functions:

1. `plane3D`: Plot a three-dimensional regression prediction with confidence intervals.
2. `twoplanes3D`: Plot a three-dimensional regression prediction with two planes, typically separated at a cut point in one of the two horizontal dimensions, with their respective confidence intervals.
3. `heatmap3D`: Auxiliary function for adding three-dimensional heatmaps to plots produced by either `plane3D` or `twoplanes3D`. These heatmaps show the joint frequency/density distribution of the model predictors represented on the horizontal axes of the plots.
}

\examples{
\dontrun{
example(regplane3D)
example(twoplanes3D)
example(heatmap3D)
}
}
\seealso{
\code{\link{plot3D}}

\code{\link{graphics}}
}
