% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twoplanes3D.R
\name{twoplanes3D}
\alias{twoplanes3D}
\title{twoplanes3D}
\usage{
twoplanes3D(
  z,
  x,
  y,
  z2,
  x2,
  y2,
  zlab = NULL,
  xlab = NULL,
  ylab = NULL,
  zlim = NULL,
  xlim = NULL,
  ylim = NULL,
  cis = TRUE,
  nlines = 5,
  heatmap = NULL,
  heatmap_cols = NA,
  heatmap_border = NA,
  plane0 = FALSE,
  zplane_aux = FALSE,
  zplane_aux_pos = NULL,
  zplane_aux_col = NULL,
  yplane_aux = FALSE,
  yplane_aux_pos = NULL,
  yplane_aux_col = NULL,
  xplane_aux = FALSE,
  xplane_aux_pos = NULL,
  xplane_aux_col = NULL,
  main = NULL,
  cex.main = 1,
  nticks = NULL,
  expand = 0.8,
  theta = 45,
  phi = 0,
  r = sqrt(3),
  d = 1.2
)
}
\arguments{
\item{z}{three-dimensional array of point estimates and interval values for
the first plane.}

\item{x}{A vector of \code{x}-values, matching the first dimension of
\code{z} for the first plane.}

\item{y}{A vector of \code{y}-values, matching the second dimension of
\code{z} for the first plane.}

\item{z2}{three-dimensional array of point estimates and interval values for
the second plane.}

\item{x2}{A vector of \code{x}-values, matching the first dimension of
\code{z} for the second plane.}

\item{y2}{A vector of \code{y}-values, matching the second dimension of
\code{z} for the second plane.}

\item{zlab}{Label for the vertical \code{z}-axis.}

\item{xlab}{Label for the horizontal \code{x}-axis.}

\item{ylab}{Label for the horizontal \code{y}-axis.}

\item{zlim}{Range of the \code{z}-axis.}

\item{xlim}{Range of the \code{x}-axis.}

\item{ylim}{Range of the \code{y}-axis.}

\item{cis}{Plot confidence intervals? Defaults to \code{TRUE}.}

\item{nlines}{Suggested number of lines for the grid display of the
regression plane. May be overridden.}

\item{heatmap}{A two-dimensional matrix indicating the joint frequency
distribution of \code{x} and \code{y}. Will displayed as a shaded histogram
below the plot.}

\item{heatmap_cols}{Override default shadings of \code{heatmap}.}

\item{heatmap_border}{Override default border color of \code{heatmap}.}

\item{plane0}{Add a a horizontal red plane at \code{z == 0}.
Defaults to \code{FALSE}.}

\item{zplane_aux}{Add an auxiliary horizontal reference place. Defaults to
\code{FALSE}.}

\item{zplane_aux_pos}{User-specified position of auxiliary horizontal
reference plane if \code{zplane_aux == TRUE}.}

\item{zplane_aux_col}{User-specified color of auxiliary horizontal reference
plane if \code{zplane_aux == TRUE}.}

\item{yplane_aux}{Add an auxiliary vertical reference place. Defaults to
\code{FALSE}.}

\item{yplane_aux_pos}{User-specified position of auxiliary vertical reference
plane if \code{yplane_aux == TRUE}.}

\item{yplane_aux_col}{User-specified color of auxiliary vertical reference
plane if \code{yplane_aux == TRUE}.}

\item{xplane_aux}{Add an auxiliary vertical reference place. Defaults to
\code{FALSE}.}

\item{xplane_aux_pos}{User-specified position of auxiliary vertical reference
plane if \code{xplane_aux == TRUE}.}

\item{xplane_aux_col}{User-specified color of auxiliary vertical reference
plane if \code{xplane_aux == TRUE}.}

\item{...}{Additional arguments passed to the plotting methods. See
\code{\link{persp3D}} and \code{\link{persp}}.}
}
\value{
Returns, as invisible, the viewing transformation matrix.
}
\description{
Plot a three-dimensional regression prediction with two planes,
typically separated at a cut point in one of the two horizontal dimensions,
and their respective confidence intervals.
}
\examples{
## ---- Estimation ----
mod <-
  lm(vote ~
       growth +
       centered_approval +
       approval_above_mean +
       growth:approval_above_mean +
       centered_approval:approval_above_mean,
     dat = us)

## ---- Prediction ----
growth_range <- c(-4, 6)
approval_range <- c(-20, 30)
approval_above_mean_vals <- c(0, 1)
growth_seq <-
  seq(growth_range[1], growth_range[2], length.out = 21L)
approval_seq <-
  seq(approval_range[1], 0, length.out = 21L)

pred <- array(NA, dim = c(21L, 21L, 2L, 3L))
for (growth in seq_along(growth_seq)) {
  for (centered_approval in seq_along(approval_seq)) {
    for (approval_above_mean in seq_along(approval_above_mean_vals)) {
      pred_tmp <- predict.lm(
        mod,
        newdata = data.frame(
          growth = growth_seq[growth],
          centered_approval = approval_seq[centered_approval] -
            min(approval_seq) *
            approval_above_mean_vals[approval_above_mean],
          approval_above_mean = approval_above_mean_vals[approval_above_mean]
        ),
        se.fit = TRUE
      )
      pred[growth, centered_approval, approval_above_mean,] <-
        c(
          pred_tmp$fit,
          pred_tmp$fit + qnorm(.025) * pred_tmp$se.fit,
          pred_tmp$fit + qnorm(.975) * pred_tmp$se.fit
        )
    }
  }
}

## ---- Plot ----
twoplanes3D(
  z = pred[, , 1, ],
  x = growth_seq,
  y = approval_seq,
  z2 = pred[, , 2, ],
  x2 = growth_seq ,
  y2 = approval_seq - min(approval_seq),
  zlab = "Predicted Vote Share",
  xlab = "Economic Growth",
  ylab = "Approval Rating \n Above & Below Average",
  zlim = c(30, 75),
  xlim = growth_range,
  ylim = c(min(approval_seq), -min(approval_seq)),
  nlines = 7,
  main = "Incumbent Vote Shares, Economic \n Growth, and Approval Ratings",
  theta = -55,
  phi = 9
)
}
\references{
\url{https://www.mzes.uni-mannheim.de/socialsciencedatalab/article/regplane3D/}
}
\seealso{
\code{\link{plot3D}}
}
