% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twoplanes3D.R
\name{twoplanes3D}
\alias{twoplanes3D}
\title{twoplanes3D}
\usage{
twoplanes3D(
  z,
  x,
  y,
  z2,
  x2,
  y2,
  zlab = NULL,
  xlab = NULL,
  ylab = NULL,
  zlim = NULL,
  xlim = NULL,
  ylim = NULL,
  xnlines = 5,
  ynlines = 5,
  cis = TRUE,
  heatmap = NULL,
  heatmap_cols = NULL,
  heatmap_border = NA,
  plane0 = FALSE,
  zplane_aux = FALSE,
  zplane_aux_pos = NULL,
  zplane_aux_col = NULL,
  yplane_aux = FALSE,
  yplane_aux_pos = NULL,
  yplane_aux_col = NULL,
  xplane_aux = FALSE,
  xplane_aux_pos = NULL,
  xplane_aux_col = NULL,
  main = NULL,
  cex.main = 1,
  expand = 0.8,
  theta = 45,
  phi = 0,
  r = sqrt(3),
  d = 1.2
)
}
\arguments{
\item{z}{three-dimensional array of point estimates and interval values for
the first plane.}

\item{x}{A vector of \code{x}-values, matching the first dimension of
\code{z} for the first plane.}

\item{y}{A vector of \code{y}-values, matching the second dimension of
\code{z} for the first plane.}

\item{z2}{three-dimensional array of point estimates and interval values for
the second plane.}

\item{x2}{A vector of \code{x}-values, matching the first dimension of
\code{z} for the second plane.}

\item{y2}{A vector of \code{y}-values, matching the second dimension of
\code{z} for the second plane.}

\item{zlab}{titles for the axes.  N.B. These must be
    character strings; expressions are not accepted.  Numbers will be
    coerced to character strings.}

\item{xlab}{titles for the axes.  N.B. These must be
    character strings; expressions are not accepted.  Numbers will be
    coerced to character strings.}

\item{ylab}{titles for the axes.  N.B. These must be
    character strings; expressions are not accepted.  Numbers will be
    coerced to character strings.}

\item{zlim}{x-, y-  and z-limits.  These should be chosen
    to cover the range of values of the surface: see \sQuote{Details}.}

\item{xlim}{x-, y-  and z-limits.  These should be chosen
    to cover the range of values of the surface: see \sQuote{Details}.}

\item{ylim}{x-, y-  and z-limits.  These should be chosen
    to cover the range of values of the surface: see \sQuote{Details}.}

\item{xnlines}{Suggested number of lines for the grid display of the
regression plane along the \code{x}-dimension. See
\code{\link{pretty_axis_inputs}}.}

\item{ynlines}{Suggested number of lines for the grid display of the
regression plane along the \code{y}-dimension. See
\code{\link{pretty_axis_inputs}}.}

\item{cis}{Plot confidence intervals? Defaults to \code{TRUE}.}

\item{heatmap}{A two-dimensional matrix indicating the joint frequency
distribution of \code{x} and \code{y}. Will displayed as a shaded histogram
below the plot.}

\item{heatmap_cols}{Override default shadings of \code{heatmap}.}

\item{heatmap_border}{Override default border color of \code{heatmap}.}

\item{plane0}{Add a a horizontal red plane at \code{z == 0}.
Defaults to \code{FALSE}.}

\item{zplane_aux}{Add an auxiliary horizontal reference place. Defaults to
\code{FALSE}.}

\item{zplane_aux_pos}{User-specified position of auxiliary horizontal
reference plane if \code{zplane_aux == TRUE}.}

\item{zplane_aux_col}{User-specified color of auxiliary horizontal reference
plane if \code{zplane_aux == TRUE}.}

\item{yplane_aux}{Add an auxiliary vertical reference place. Defaults to
\code{FALSE}.}

\item{yplane_aux_pos}{User-specified position of auxiliary vertical reference
plane if \code{yplane_aux == TRUE}.}

\item{yplane_aux_col}{User-specified color of auxiliary vertical reference
plane if \code{yplane_aux == TRUE}.}

\item{xplane_aux}{Add an auxiliary vertical reference place. Defaults to
\code{FALSE}.}

\item{xplane_aux_pos}{User-specified position of auxiliary vertical reference
plane if \code{xplane_aux == TRUE}.}

\item{xplane_aux_col}{User-specified color of auxiliary vertical reference
plane if \code{xplane_aux == TRUE}.}

\item{main}{main and sub title, as for \code{\link[graphics]{title}}.}

\item{cex.main}{The magnification to be used for main titles relative to the
current setting of cex as in \code{\link[graphics]{par}}.}

\item{expand}{a expansion factor applied to the \code{z}
    coordinates. Often used with \code{0 < expand < 1} to shrink the
    plotting box in the \code{z} direction.}

\item{theta}{angles defining the viewing direction.
    \code{theta} gives the azimuthal direction and \code{phi}
    the colatitude.}

\item{phi}{angles defining the viewing direction.
    \code{theta} gives the azimuthal direction and \code{phi}
    the colatitude.}

\item{r}{the distance of the eyepoint from the centre of the plotting box.}

\item{d}{a value which can be used to vary the strength of
    the perspective transformation.  Values of \code{d} greater
    than 1 will lessen the perspective effect and values less
    and 1 will exaggerate it.}
}
\value{
Returns, as invisible, the viewing transformation matrix.
}
\description{
Plot a three-dimensional regression prediction with two planes,
typically separated at a cut point in one of the two horizontal dimensions,
and their respective confidence intervals.
}
\examples{
## ---- Estimation ----
mod <- lm(
  vote ~
    growth +
    centered_approval +
    approval_above_mean +
    growth:approval_above_mean +
    centered_approval:approval_above_mean,
  dat = us
)

## ---- Axis inputs ----
## Cut point
cut_point <- 0
approval_above_mean_vals <- c(0, 1)

## Growth
growth_axis <- pretty_axis_inputs(
  axis_range = range(us$growth),
  base = 2,
  nlines_suggest = 6L,
  multiply = 4
)

## Approval
approval_axis <- pretty_axis_inputs(
  axis_range = c(min(us$centered_approval), cut_point),
  base = 10,
  nlines_suggest = 3L,
  multiply = 4
)

## ---- Prediction ----
pred <-
  array(NA, dim = c(length(growth_axis$seq), length(approval_axis$seq), 2L, 3L))
for (growth in seq_along(growth_axis$seq)) {
  for (centered_approval in seq_along(approval_axis$seq)) {
    for (approval_above_mean in seq_along(approval_above_mean_vals)) {
      pred_tmp <- predict.lm(
        mod,
        newdata = data.frame(
          growth = growth_axis$seq[growth],
          centered_approval = approval_axis$seq[centered_approval] -
            min(approval_axis$seq) *
            approval_above_mean_vals[approval_above_mean],
          approval_above_mean = approval_above_mean_vals[approval_above_mean]
        ),
        se.fit = TRUE
      )

      pred[growth, centered_approval, approval_above_mean,] <-
        c(
          pred_tmp$fit,
          pred_tmp$fit + qnorm(.025) * pred_tmp$se.fit,
          pred_tmp$fit + qnorm(.975) * pred_tmp$se.fit
        )
    }
  }
}

## ---- Plot ----
twoplanes3D(
  z = pred[, , 1,],
  x = growth_axis$seq,
  y = approval_axis$seq,
  z2 = pred[, , 2,],
  x2 = growth_axis$seq,
  y2 = approval_axis$seq - min(approval_axis$seq),
  zlim = c(35, 70),
  xlim = growth_axis$range,
  ylim = c(min(approval_axis$seq),-min(approval_axis$seq)),
  zlab = "Predicted Vote Share",
  xlab = "Economic Growth",
  ylab = "Approval Rating \n Above & Below Average",
  cis = TRUE,
  xnlines = growth_axis$nlines,
  ynlines = approval_axis$nlines,
  main = "Incumbent Vote Shares, Economic \n Growth, and Approval Ratings",
  theta = -55,
  phi = 9
)
}
\references{
\url{https://www.mzes.uni-mannheim.de/socialsciencedatalab/article/regplane3D/}
}
\seealso{
\code{\link{plot3D}}
}
